+++ lib/mongoid/config.rb
@@ -15,7 +15,7 @@ module Mongoid
 
     delegate :logger=, to: ::Mongoid
     delegate :logger, to: ::Mongoid
-    
+
     LOCK = Mutex.new
 
     option :allow_dynamic_fields, default: true